Fixed the kernel calling with explictly casted arguments. This is needed for newer HIP compiler. And added the empty namespaces which are also supported

+++ caffe2/operators/batch_gather_ops_hip.cc
@@ -63,8 +63,8 @@ bool BatchGatherOp<HIPContext>::DoRunWithType() {
       idxs,
       M,
       N,
-      data_batch_size,
-      gathered_batch_size,
+      static_cast<const int>(data_batch_size),
+      static_cast<const int>(gathered_batch_size),
       block_size);
   return true;
 }
@@ -136,8 +136,8 @@ bool BatchGatherGradientOp<HIPContext>::DoRunWithType2() {
       idxs,
       M,
       N,
-      data_batch_size,
-      gathered_batch_size,
+      static_cast<const int>(data_batch_size),
+      static_cast<const int>(gathered_batch_size),
       block_size);
 
   return true;
